fix: update CLI import paths to match new command directory structure
The CLI commands were reorganized from individual files to subdirectories
with index.ts files, but the CLI import paths weren't updated accordingly.

- Update imports in src/index.ts to use new paths:
  - commands/dev.js → commands/dev/index.js
  - commands/compile.js → commands/compile/index.js
  - commands/clean.js → commands/clean/index.js
  - commands/bundle.js → commands/bundle/index.js
  - commands/deploy.js → commands/deploy/index.js

This fixes the CLI build which was failing with module resolution errors.
