+++ core/cli-core/src/commands/clean/index.ts
