+++ core/cli-core/src/commands/compile/index.ts
